feat: add chibihash v2 support
