+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "chibihash"
-version = "0.3.1"
+version = "0.4.0"
 edition = "2021"
 license = "MIT"
 authors = ["Ville Vesilehto <ville@vesilehto.fi>"]
